fix op_global_set

+++ src/test/test_wasm_instructions/op_global_set.rs
@@ -6,12 +6,12 @@ fn test_global_set() {
     let textual_repr = r#"
         (module
             (global $global_i32 (mut i32) (i32.const 10))
-            (global $global_i64 (mut i32) (i32.const 10))
+            (global $global_i64 (mut i64) (i64.const 10))
 
             (func (export "test")
                 (i32.const 0)
                 (global.set $global_i32)
-                (i32.const 0)
+                (i64.const 0)
                 (global.set $global_i64)
             )
         )
